feat: Add logging to ai_benchmark execution
The code changes add logging statements using the `loguru` library to track the execution of the `ai_benchmark` module. This will provide more visibility into the benchmarking process and help with debugging if any issues arise.

Based on the recent repository commits, the established convention for chore-related commits is to start with the prefix "chore:". However, since this change adds a new feature (logging), the commit message should start with the prefix "feat:".

Please note that the commit message should not include any meta information like issue references, tags, or author names.

Signed-off-by: haitwang-cloud <haitao_wht@outlook.com>